Add files via upload
The code has been included in the given repository. It shows detailed steps used for splitting the data in training and testing model. 
Logistic Regression has been used to find the result after putting the details which were used to train the model. I have used logistic regression or 'sigmoid function' to take input as independent variables and produce a probability between 0 and 1. 
The training of this model has shown an accuracy of 0.83.